+++ Tienda/src/main/java/Model/Producto.java
@@ -2,14 +2,14 @@ package Model;
 
 public class Producto {
     private int id_Producto;
-    private int codigo;
+    private long codigo;
     private String nombre;
     private int cantidad;
     private double precio_venta;
     private double costo_compra;
     private int activo;
 
-    public Producto(int id_Producto, int codigo, String nombre, int cantidad, double precio_venta, double costo_compra, int activo) {
+    public Producto(int id_Producto, long codigo, String nombre, int cantidad, double precio_venta, double costo_compra, int activo) {
         this.id_Producto = id_Producto;
         this.codigo = codigo;
         this.nombre = nombre;
@@ -19,14 +19,13 @@ public class Producto {
         this.activo = activo;
     }
 
-    public Producto(int codigo, String nombre, int cantidad, double precio_venta, double costo_compra) {
+    public Producto(long codigo, String nombre, int cantidad, double precio_venta, double costo_compra) {
         this.codigo = codigo;
         this.nombre = nombre;
         this.cantidad = cantidad;
         this.precio_venta = precio_venta;
         this.costo_compra = costo_compra;
     }
-    
 
     public int getId_Producto() {
         return id_Producto;
@@ -36,11 +35,11 @@ public class Producto {
         this.id_Producto = id_Producto;
     }
 
-    public int getCodigo() {
+    public long getCodigo() {
         return codigo;
     }
 
-    public void setCodigo(int codigo) {
+    public void setCodigo(long codigo) {
         this.codigo = codigo;
     }
 
@@ -83,4 +82,5 @@ public class Producto {
     public void setActivo(int activo) {
         this.activo = activo;
     }
+
 }
